Seperated Side Bar Generation Code
